[BE] refactor: 기존의 MapBuffer를 사용하도록 리팩토링  #763
