+++ backend/app/src/main/java/com/onair/hearit/app/playinghistory/infrastructure/buffer/PlayingHistoryMapBuffer.java
@@ -1,23 +1,20 @@
 package com.onair.hearit.app.playinghistory.infrastructure.buffer;
 
 import com.onair.hearit.app.exception.custom.BufferRequestException;
-import com.onair.hearit.core.domain.Hearit;
+import com.onair.hearit.app.playinghistory.infrastructure.buffer.converter.PlayingHistoryConverter;
+import com.onair.hearit.app.playinghistory.infrastructure.buffer.converter.PlayingHistoryConverter.PlayKey;
+import com.onair.hearit.app.playinghistory.infrastructure.buffer.converter.PlayingHistoryConverter.PlayValue;
 import com.onair.hearit.core.domain.PlayingHistory;
 import com.onair.hearit.core.infrastructure.jdbc.PlayingHistoryCommandRepository;
-import com.onair.hearit.core.infrastructure.jpa.HearitRepository;
-import jakarta.annotation.PreDestroy;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.stereotype.Component;
 
 @Slf4j
+@Component
 @RequiredArgsConstructor
 public class PlayingHistoryMapBuffer implements PlayingHistoryBuffer {
 
@@ -25,15 +22,34 @@ public class PlayingHistoryMapBuffer implements PlayingHistoryBuffer {
 
     private final Map<PlayKey, PlayValue> cache = new ConcurrentHashMap<>();
     private final PlayingHistoryCommandRepository playingHistoryCommandRepository;
-    private final HearitRepository hearitRepository;
+    private final PlayingHistoryConverter converter;
+
+//    @Override
+//    public void add(PlayingHistory playingHistory, long clientEventTime) {
+//        PlayKey key = new PlayKey(playingHistory.getUserUuid(), playingHistory.getHearitId());
+//        cache.compute(key, (k, existing) -> {
+//            if (existing == null) {
+//                validateBufferSize(k);
+//            }
+//            PlayValue incoming = PlayValue.from(playingHistory, clientEventTime);
+//            if (existing != null && existing.isMoreRecentThan(incoming)) {
+//                return existing;
+//            }
+//            return incoming;
+//        });
+//    }
 
     @Override
     public void add(PlayingHistory playingHistory, long clientEventTime) {
+        validateClientEventTime(clientEventTime);
         PlayKey key = new PlayKey(playingHistory.getUserUuid(), playingHistory.getHearitId());
+
+        if (!cache.containsKey(key) && cache.size() >= BUFFER_SIZE) {
+            log.error("Fallback 캐시 용량 초과: {}", cache.size());
+            throw new BufferRequestException("버퍼 용량 초과로 인해 재생 기록 저장할 수 없습니다.");
+        }
+
         cache.compute(key, (k, existing) -> {
-            if (existing == null) {
-                validateBufferSize(k);
-            }
             PlayValue incoming = PlayValue.from(playingHistory, clientEventTime);
             if (existing != null && existing.isMoreRecentThan(incoming)) {
                 return existing;
@@ -54,20 +70,18 @@ public class PlayingHistoryMapBuffer implements PlayingHistoryBuffer {
     }
 
     @Override
-    @Transactional
-    @Scheduled(fixedDelay = 1_000)
     public void flush() {
         Map<PlayKey, PlayValue> snapshot = createSnapshotAndRemoveFromCache();
         if (snapshot.isEmpty()) {
             return;
         }
-
         try {
-            List<PlayingHistory> histories = buildHistoriesFromSnapshot(snapshot);
+            List<PlayingHistory> histories = converter.toPlayingHistories(snapshot.values());
             playingHistoryCommandRepository.bulkInsert(histories);
         } catch (Exception e) {
             rollbackSnapshot(snapshot);
             log.error("재생 기록 flush 실패, 롤백 수행. snapshot size: {}", snapshot.size(), e);
+            throw new RuntimeException("Local storage flush 실패 ", e);
         }
     }
 
@@ -81,57 +95,26 @@ public class PlayingHistoryMapBuffer implements PlayingHistoryBuffer {
         return snapshot;
     }
 
-    private List<PlayingHistory> buildHistoriesFromSnapshot(Map<PlayKey, PlayValue> snapshot) {
-        Set<Long> hearitIds = snapshot.values().stream()
-                .map(PlayValue::hearitId)
-                .collect(Collectors.toSet());
-        Map<Long, Hearit> hearitMap = hearitRepository.findAllById(hearitIds)
-                .stream()
-                .collect(Collectors.toMap(Hearit::getId, h -> h));
-        return snapshot.values().stream().map(v -> new PlayingHistory(
-                        v.userUuid(),
-                        hearitMap.get(v.hearitId()),
-                        v.lastPlayTime()
-                ))
-                .toList();
-    }
-
     private void rollbackSnapshot(Map<PlayKey, PlayValue> snapshot) {
-        snapshot.forEach((key, newValue) ->
-                cache.merge(key, newValue, (oldValue, incomingValue) -> {
-                    if (oldValue.isMoreRecentThan(incomingValue)) {
-                        return oldValue;
-                    }
-                    return incomingValue;
-                })
-        );
-    }
+        try {
+            snapshot.forEach((key, newValue) ->
+                    cache.merge(key, newValue, (oldValue, incomingValue) -> {
 
-    @PreDestroy
-    public void shutdown() {
-        flush();
-    }
-
-    private record PlayKey(String userUuid, long hearitId) {
-    }
-
-    private record PlayValue(String userUuid, long hearitId, long lastPlayTime, long clientEventTime) {
-
-        static PlayValue from(PlayingHistory history, long clientEventTime) {
-            return new PlayValue(
-                    history.getUserUuid(),
-                    history.getHearitId(),
-                    history.getLastPlayTime(),
-                    clientEventTime
+                        if (oldValue.isMoreRecentThan(incomingValue)) {
+                            return oldValue;
+                        }
+                        return incomingValue;
+                    })
             );
-        }
-
-        boolean isMoreRecentThan(PlayValue other) {
-            return other != null && this.clientEventTime > other.clientEventTime();
+        } catch (Exception e) {
+            log.error("롤백도 실패, 데이터 손실 가능성. size: {}", snapshot.size(), e);
+            // 메트릭 기록 또는 알림 필요
         }
     }
 
-    public Map<PlayKey, PlayValue> getCache() {
-        return Collections.unmodifiableMap(cache);
+    private void validateClientEventTime(long clientEventTime) {
+        if (clientEventTime <= 0) {
+            throw new IllegalArgumentException("clientEventTime must be positive: " + clientEventTime);
+        }
     }
 }
